+++ src/VirtoCommerce.PushMessages.Web/App/src/modules/push-messages/components/index.ts
@@ -1,1 +1,2 @@
 export * from "./widgets";
+export { default as BaseListBlade } from "./BaseListBlade.vue";
