+++ src/VirtoCommerce.PushMessages.Web/App/src/modules/push-messages/composables/index.ts
@@ -5,3 +5,4 @@ export { useTrackingList } from "./useTrackingList";
 export { useSentList } from "./useSentList";
 export { useMessageDetails } from "./useMessageDetails";
 export { useRecipientList } from "./useRecipientList";
+export { useBaseList } from "./useBaseList";
