+++ src/VirtoCommerce.PushMessages.Web/App/src/modules/push-messages/utils/index.ts
@@ -1,0 +1,1 @@
+export { useMessageListColumns } from "./columns";
